requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==25.3.0
blinker==1.8.2
ccxt==4.5.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
cryptography==45.0.6
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.0.3
frozenlist==1.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.1
MarkupSafe==2.1.5
multidict==6.1.0
numpy==1.24.4
openai==1.99.9
pandas==2.0.3
propcache==0.2.0
pycares==4.4.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
ta==0.11.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
Werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
